+++ MusicKit/Frameworks/SndKit/SndAudioProcessorFlanger.m
@@ -64,7 +64,7 @@
 // setParameter
 //////////////////////////////////////////////////////////////////////////////
 
-- (void) setParam: (const int) index toValue: (const float) v
+- (void) setParam: (const NSInteger) index toValue: (const float) v
 {
   float value = v > 1.0 ? 1.0 : (v < 0.0 ? 0.0 : v);
   switch (index)  {
@@ -81,7 +81,7 @@
 // paramValue
 //////////////////////////////////////////////////////////////////////////////
 
-- (float) paramValue: (const int) index
+- (float) paramValue: (const NSInteger) index
 {
   float r = 0.0f;
   switch (index)
@@ -100,7 +100,7 @@
 // paramName
 //////////////////////////////////////////////////////////////////////////////
 
-- (NSString*) paramName: (const int) index
+- (NSString*) paramName: (const NSInteger) index
 {
   NSString *r = nil;
 	switch (index)
@@ -119,7 +119,7 @@
 // paramDisplay
 //////////////////////////////////////////////////////////////////////////////
 
-- (NSString*) paramDisplay: (const int) index
+- (NSString*) paramDisplay: (const NSInteger) index
 {
   NSString *r = nil;
 	switch (index)
@@ -138,7 +138,7 @@
 // paramLabel
 //////////////////////////////////////////////////////////////////////////////
 
-- (NSString*) paramLabel:(const int) index
+- (NSString*) paramLabel:(const NSInteger) index
 {
   NSString *r = nil;
 	switch (index)
